admin/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-mysql==4.12.0
djangorestframework==3.14.0
mysqlclient==2.2.1
pika==1.3.2
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
